refactor(local-service): split execution trace helpers
